Prefactor: Extract some common components.
