+++ python/shark_turbine/kernel/compiler/base.py
@@ -1,0 +1,6 @@
+class CodegenError(Exception):
+    ...
+
+
+class ValidationError(CodegenError):
+    ...
